move errors to error.py
